+++ ChangeLog.md
@@ -2,6 +2,11 @@
 
 ## OpenRaider (0.1.3) xythobuz <xythobuz@xythobuz.de>
 
+    [ 20140503 ]
+    * Started big World-Data refactoring.
+      - Created Sprite class with drawing code from Render
+      - Created Room, Portal, Box, Sector, StaticModel classes
+
 	[ 20140501 ]
 	* Fix compiling with gcc on Arch Linux
 	- carstene1ns <dev@f4ke.de>